+++ lib/moonscript/util.lua
@@ -1,4 +1,4 @@
-local fenv=require('fenv_kostyl')
+local fenv=require('fenv_hack')
 local concat
 concat = table.concat
 local unpack = unpack or table.unpack
